% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getGRexample.R
\name{getGRexample}
\alias{getGRexample}
\title{getGRexample}
\usage{

  getGRexample(url = "https://github.com/hilldr/gene_retriever/raw/getGRexample/example_normout.tar.gz")
}
\arguments{
\item{url}{Specifies the URL path of the file to download.}
}
\value{
Downloads an example dataset in the working directory
}
\description{
This function downloads and unpacks an example dataset in the working directory. See ?GeneRetriever for additional examples.
}
\examples{
getGRexample()
}

