% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSRexample.R
\name{getSRexample}
\alias{getSRexample}
\title{getSRexample}
\usage{

  getSRexample(url = "https://github.com/hilldr/SeqRetriever/raw/master/example_normout.tar.gz")
}
\arguments{
\item{url}{Specifies the URL path of the file to download.}
}
\value{
Downloads an example dataset in the working directory
}
\description{
This function downloads and unpacks an example dataset in the working directory. See ?SeqRetriever for additional examples.
}
\examples{
getSRexample()
}

