% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqANOVA.R
\name{SeqANOVA}
\alias{SeqANOVA}
\title{SeqANOVA}
\usage{
SeqANOVA(df = df, p.adjust.method = "bonferroni", ...)
}
\arguments{
\item{df}{The dataframe generated by SeqDataframe}

\item{p.adjust.method}{Correction method for ANOVA p-value. Accepted options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". See ?p.adjust for discussion of methods. Default = "bonferroni"}
}
\value{
dataframe containing FPKM values and ANOVA p-value and adjusted ANOVA p-value for all groups.
}
\description{
This function performs Analysis of variance (ANOVA) on SeqDataframe() format dataframes and returns a P-value. Groups are selected automatically based on name e.g. group1_0, group1_1, group1_2, group2_0, group2_1, group2_2,group3_0, group3_1, group3_2 would perform ANOVA on group1, group2, and group3. Returns p-value (anova.p) and adjusted p-vale (anova.p.adj)
}
\examples{
results <- SeqANOVA(SeqDataframe(dir="./norm_out"))

}

