% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqBoxplot.R
\name{SeqBoxplot}
\alias{SeqBoxplot}
\title{SeqBoxplot}
\usage{
SeqBoxplot(df, nrow = 3, scales = "free_y", size = 5)
}
\arguments{
\item{df}{Name of the SeqGenes format dataframe to plot}

\item{nrow}{The number of rows in boxplot array. Default 3}

\item{scales}{Determines how scales are set: "fixed", free ("free"), or free in one dimension ("free_x", "free_y"). Default is "free_y". See ggplot2 facet_wrap() documentation.}

\item{size}{Integer. Point size points mapped by geom_point(). Set to 0 for box and whiskers plots only. Default 5}
}
\value{
ggplot 2 facet_wrap() opject
}
\description{
This function searches cuffnorm format gene expression data for user specified genes and generates a ggplot2 format facet grid with normalized FPKM boxplots
}
\examples{
getSRexample() # Downloads and unpacks example dataset in working directory
testdf <- SeqDataframe(dir = "./norm_out")
genes <- SeqGenes(gene.names = c("OR4F5","SAMD11","AJAP1","SKI","ESPN", "CNKSR1"), df = testdf)
plot <- SeqBoxplot(genes)
print(plot)
}

