% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqDataframe.R
\name{SeqDataframe}
\alias{SeqDataframe}
\title{SeqDataframe}
\usage{
SeqDataframe(dir = "./")
}
\arguments{
\item{dir}{The directory containing CuffNorm format output files.}
}
\value{
Normalized FPKM dataframe of FPKM counts and gene metadata accross all samples.
}
\description{
This function accepts cuffnorm format fpkm counts and returns a formatted R dataframe
}
\examples{
getSRexample() # Downloads and unpacks example dataset in working directory
SeqDataframe(dir="./norm_out")
}

