% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqGenes.R
\name{SeqGenes}
\alias{SeqGenes}
\title{SeqGenes}
\usage{
SeqGenes(df, gene.names, csv = FALSE, csv.out = "SeqGenesOutput.csv")
}
\arguments{
\item{df}{SeqDataframe object to search}

\item{gene.names}{This first argument is a vector of gene names (as string). Gene names MUST be NCBI Genbank format.}

\item{csv}{Boolean operator controlling csv output. TRUE returns csv file with filename as specified by 'csv.out'. Default FALSE}

\item{csv.out}{Name and location of the CSV file output. Default "SRoutput.csv"}
}
\value{
dataframe
}
\description{
This function subsets a SeqDataframe based on a user-supplied gene list
}
\examples{
getSRexample() # Downloads and unpacks example dataset in working directory
testdf <- SeqDataframe(dir = "./norm_out")
genes <- SeqGenes(gene.names = c("OR4F5","SAMD11","AJAP1","SKI","ESPN", "CNKSR1"), df = testdf)
}

