% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeqHeatmap.R
\name{SeqHeatmap}
\alias{SeqHeatmap}
\title{SeqHeatmap}
\usage{
SeqHeatmap(df, hm.name = "SRheatmap.png", w = 8, h = 11, cellwidth = 15,
  cellheight = 15)
}
\arguments{
\item{hm.name}{Name of heatmap output. Default is "gr_heatmap.pdf"}

\item{cellwidth}{Heatmap cell width in px. Default 30}

\item{cellheight}{Heatmap cell height in px. Default 30}
}
\value{
Normalized FPKM matrix containing the specified subset of genes accross all samples. Additional options will plot expression of individual genes as box plots and/or a heatmap with hierarchical clustering
}
\description{
This function searches cuffnorm format gene expression data for user specified genes and generates a normalized FPKM table and box plots and/or a heatmap with hierarchical clustering.
}
\examples{
getSRexample() # Downloads and unpacks example dataset in working directory
SeqRetriever(gene.names=c("OR4F5","SAMD11","AJAP1","SKI","ESPN", "CNKSR1"),nrow=3,dir="./norm_out", boxplot = TRUE, heatmap = TRUE)
}

