% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/SeqRetriever.R
\name{SeqRetriever}
\alias{SeqRetriever}
\title{SeqRetriever}
\usage{
SeqRetriever(gene.names, nrow = 3, dir = "./", csv.out = "SRoutput.csv",
  bp.name = "SRoutput.png", w = 8, h = 11, boxplot = TRUE,
  heatmap = TRUE, hm.name = "SRheatmap.png", cellwidth = 15,
  cellheight = 15)
}
\arguments{
\item{gene.names}{This first argument is a vector of gene names (as string). Genenames MUST be NCBI Genbank format.}

\item{nrow}{The number of rows in boxplot array. Default 3}

\item{dir}{The directory containing CuffNorm format output files.}

\item{csv.out}{Name and location of the CSV file output. Default "gr_output.csv"}

\item{bp.name}{Name of boxplot pdf output. Default "gr_output.pdf"}

\item{w}{Width in inches of the boxplot output. Default 8}

\item{h}{Height in inches of the boxplot output. Default 11}

\item{boxplot}{Boolean operator controlling boxplot output. TRUE returns boxplot. FALSE bypasses boxplot generation. Default TRUE}

\item{heatmap}{Boolean operator controlling heatmap output. TRUE returns heatmap plot. FALSE bypasses heatmap geneation. Default is TRUE}

\item{hm.name}{Name of heatmap output. Default is "gr_heatmap.pdf"}

\item{cellwidth}{Heatmap cell width in px. Default 30}

\item{cellheight}{Heatmap cell height in px. Default 30}
}
\value{
Normalized FPKM matrix containing the specified subset of genes accross all samples. Additional options will plot expression of individual genes as box plots and/or a heatmap with hierarchical clustering
}
\description{
This function searches cuffnorm format gene expression data for user specified genes and generates a normalized FPKM table and box plots and/or a heatmap with hierarchical clustering.
}
\examples{
getSRexample() # Downloads and unpacks example dataset in working directory
GeneRetriever(gene.names=c("OR4F5","SAMD11","AJAP1","SKI","ESPN", "CNKSR1"),nrow=3,dir="./norm_out", boxplot = TRUE, heatmap = TRUE)
}

