% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqStatSubset.R
\name{SeqStatSubset}
\alias{SeqStatSubset}
\title{SeqStatSubset}
\usage{
SeqStatSubset(df = df, method = "t.test", limit = 0.05,
  p.adjust = FALSE, p.adjust.method = "bonferroni", group1 = "HLO",
  group2 = "ES")
}
\arguments{
\item{df}{The dataframe generated by SeqDataframe}

\item{method}{Statistical test to be performed. Accepted options are "t.test" and "anova". Option "t.test' requires specification of group1 and group2; these variables are ignored if "anova" is selected. Default = "t.test"}

\item{limit}{A value representing the threshold of statistical significance, or maximum P-value considered to represent a significant difference. Default is set at 0.05.}

\item{p.adjust}{Boolean operator to determine the use of multiple testing correction correction when subsetting based on P-value. Default is FALSE, meaning that no multiple testing correction is performed.}

\item{p.adjust.method}{Multiple testing correction method for  p-value. Accepted options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". See ?p.adjust for discussion of methods. Default = "bonferroni"}

\item{group1}{A string unique to the columns in 'data' that contain the first data set to be compared}

\item{group2}{A string unique to the columns in 'data' that contain the second data set to be compared}
}
\description{
This function subsets a SeqDataframe according to the results of a two-tailed unpaired Student's t-test and returns a dataframe containing only the genes and FPKM values meeting the statistical significance criteria
}
\examples{
data <- SeqDataframe(dir="./norm_out")
subset <- SeqStatSubset(data, limit= 0.001, group1 = "ES", group2 = "HLO")
}

