% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqStudents.R
\name{SeqStudents}
\alias{SeqStudents}
\title{SeqStudents}
\usage{
SeqStudents(data, group1 = "ES", group2 = "HLO",
  p.adjust.method = "bonferroni")
}
\arguments{
\item{data}{The dataframe generated by SeqDataframe}

\item{group1}{A string unique to the columns in 'data' that contain the first data set to be compared}

\item{group2}{A string unique to the columns in 'data' that contain the second data set to be compared}

\item{p.adjust.method}{Correction method for ANOVA p-value. Accepted options are "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". See ?p.adjust for discussion of methods. Default = "bonferroni"}
}
\description{
This function performs a two-tailed Student's t-test on two groups specified by the user. The preferred input is a dataframe generated using SeqDataframe().
}
\examples{
results <- SeqStudents(SeqDataframe(dir="./norm_out"), group1 = "ES", group2 = "HLO")
}

