% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqBoxplot.R
\name{SeqBoxplot}
\alias{SeqBoxplot}
\title{SeqBoxplot}
\usage{
SeqBoxplot(gene.names, nrow = 3, dir = "./", csv.out = "SRoutput.csv",
  scales = "free_y")
}
\arguments{
\item{gene.names}{This first argument is a vector of gene names (as string). Genenames MUST be NCBI Genbank format.}

\item{nrow}{The number of rows in boxplot array. Default 3}

\item{dir}{The directory containing CuffNorm format output files.}

\item{csv.out}{Name and location of the CSV file output. Default "SRoutput.csv"}

\item{scales}{Determines how scales are set: "fixed", free ("free"), or free in one dimension ("free_x", "free_y"). Default is "free_y". See ggplot2 facet_wrap() documentation.}

\item{csv}{Boolean operator controlling csv output. TRUE returns csv file with filename as specified by 'csv.out'. Default FALSE}
}
\value{
ggplot 2 facet_wrap() opject
}
\description{
This function searches cuffnorm format gene expression data for user specified genes and generates a ggplot2 format facet grid with normalized FPKM boxplots
}
\examples{
getSRexample() # Downloads and unpacks example dataset in working directory
example <- SeqBoxplot(gene.names=c("OR4F5","SAMD11","AJAP1","SKI","ESPN", "CNKSR1"),nrow=3,dir="./norm_out")
print(example)
}

