% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqStatSubset.R
\name{SeqStatSubset}
\alias{SeqStatSubset}
\title{SeqStatSubset}
\usage{
SeqStatSubset(data, limit = 0.05, Bonferroni = FALSE, group1 = "HLO",
  group2 = "ES")
}
\arguments{
\item{data}{The dataframe generated by SeqDataframe}

\item{limit}{A value representing the threshold of statistical significance, or maximum P-value considered to represent a significant difference. Default is set at 0.05.}

\item{Bonferroni}{A boolean operator to determine the use of Bonferroni correction when subsetting based on P-value. Default is FALSE, meaning that no Bonferroni correction is performed.}

\item{group1}{A string unique to the columns in 'data' that contain the first data set to be compared}

\item{group2}{A string unique to the columns in 'data' that contain the second data set to be compared}
}
\description{
This function subsets a SeqDataframe according to the results of a two-tailed unpaired Student's t-test and returns a dataframe containing only the genes and FPKM values meeting the statistical significance criteria
}
\examples{
data <- SeqDataframe(dir="./norm_out")
subset <- SeqStatSubset(data, limit= 0.001, group1 = "ES", group2 = "HLO")
}

