% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SeqStudents.R
\name{SeqStudents}
\alias{SeqStudents}
\title{SeqStudents}
\usage{
SeqStudents(data, group1 = "ES", group2 = "HLO")
}
\arguments{
\item{data}{The dataframe generated by SeqDataframe}

\item{group1}{A string unique to the columns in 'data' that contain the first data set to be compared}

\item{group2}{A string unique to the columns in 'data' that contain the second data set to be compared}
}
\description{
This function performs a two-tailed Student's t-test on two groups specified by the user. The preferred input is a dataframe generated using SeqDataframe().
}
\examples{
results <- SeqStudents(SeqDataframe(dir="./norm_out"), group1 = "ES", group2 = "HLO")
}

